requirements.txt
pydantic==2.12.5
PyYAML==6.0.3
tinytuya==1.17.4
